+++ BotDeScans.App/Features/Titles/+Dependencies.cs
@@ -1,9 +1,8 @@
-﻿using BotDeScans.App.Features.Titles.Create;
-using BotDeScans.App.Features.Titles.Update;
-using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.Extensions.DependencyInjection;
 using Remora.Commands.Extensions;
 using Remora.Discord.Interactivity.Extensions;
 using System.Diagnostics.CodeAnalysis;
+
 namespace BotDeScans.App.Features.Titles;
 
 [ExcludeFromCodeCoverage]
@@ -11,8 +10,10 @@ internal static class AddDependencies
 {
     internal static IServiceCollection AddTitleServices(this IServiceCollection services) => services
         .AddCommandTree()
-            .WithCommandGroup<TitleCommands>()
+            .WithCommandGroup<Create.Commands>()
+            .WithCommandGroup<List.Commands>()
+            .WithCommandGroup<Update.Commands>()
             .Finish()
-        .AddInteractionGroup<CreateInteractions>()
-        .AddInteractionGroup<UpdateInteractions>();
+        .AddInteractionGroup<Create.Interactions>()
+        .AddInteractionGroup<Update.Interactions>();
 }
